Fix warnings and jest trying to run non-test files by renaming them
