+++ client/web/src/codeintel/language-specs/cpp.ts
@@ -1,5 +1,5 @@
 import { cStyleComment } from './comments'
-import { FilterContext, LanguageSpec, Result, LSIFSupport } from './spec'
+import { FilterContext, LanguageSpec, Result, LSIFSupport } from './languagespec'
 import { dotToSlash, extractFromLines, filterResultsByImports, removeExtension } from './util'
 
 /**
