+++ client/web/src/codeintel/language-specs/go.ts
@@ -1,7 +1,7 @@
 import * as path from 'path'
 
 import { slashPattern } from './comments'
-import { FilterContext, LanguageSpec, Result, LSIFSupport } from './spec'
+import { FilterContext, LanguageSpec, Result, LSIFSupport } from './languagespec'
 import { extractFromLines, filterResults } from './util'
 
 /**
