+++ client/web/src/codeintel/language-specs/java.test.ts
@@ -1,7 +1,7 @@
 import * as assert from 'assert'
 
 import { javaSpec } from './java'
-import { nilFilterContext, nilResult } from './spec.test'
+import { nilFilterContext, nilResult } from './testutils'
 
 const fileContent = `
 package com.sourcegraph.test.codeintel;
