+++ client/web/src/codeintel/language-specs/java.ts
@@ -1,7 +1,7 @@
 import * as path from 'path'
 
 import { javaStyleComment } from './comments'
-import { FilterContext, LanguageSpec, Result, LSIFSupport } from './spec'
+import { FilterContext, LanguageSpec, Result, LSIFSupport } from './languagespec'
 import { extractFromLines, filterResultsByImports, slashToDot } from './util'
 
 /**
