+++ client/web/src/codeintel/language-specs/python.test.ts
@@ -1,7 +1,7 @@
 import * as assert from 'assert'
 
 import { pythonSpec, relativeImportPath } from './python'
-import { nilFilterContext, nilResult } from './spec.test'
+import { nilFilterContext, nilResult } from './testutils'
 
 const fileContent = `
 import .bar
