+++ client/web/src/codeintel/language-specs/python.ts
@@ -1,7 +1,7 @@
 import * as path from 'path'
 
 import { pythonStyleComment } from './comments'
-import { FilterContext, LanguageSpec, Result } from './spec'
+import { FilterContext, LanguageSpec, Result } from './languagespec'
 import { extractFromLines, filterResultsByImports, removeExtension } from './util'
 
 /**
