+++ client/web/src/codeintel/language-specs/typescript.ts
@@ -1,7 +1,7 @@
 import * as path from 'path'
 
 import { cStyleComment } from './comments'
-import { FilterContext, LanguageSpec, Result, LSIFSupport } from './spec'
+import { FilterContext, LanguageSpec, Result, LSIFSupport } from './languagespec'
 import { extractFromLines, filterResultsByImports, removeExtension } from './util'
 
 /**
